Address some of Reynold's minor comments

+++ core/src/main/scala/org/apache/spark/executor/Executor.scala
@@ -189,7 +189,7 @@ private[spark] class Executor(
 
         // Run the actual task and measure its runtime.
         taskStart = System.currentTimeMillis()
-        val value = task.run(taskId = taskId.toLong, attemptId = attemptId.toLong)
+        val value = task.run(taskId = taskId, attemptId = attemptId)
         val taskFinish = System.currentTimeMillis()
 
         // If the task has been killed, let's fail it.
